refactor : LadderGameResult dto 클래스 presentation 레이어로 이동
