+++ src/main/java/ladder/view/OutputView.java
@@ -1,7 +1,7 @@
 package ladder.view;
 
 import ladder.controller.dto.LadderGenerationResult;
-import ladder.service.dto.LadderGameResult;
+import ladder.controller.dto.LadderGameResult;
 import ladder.controller.dto.LadderGameResponse;
 import ladder.controller.dto.LadderLine;
