+++ src/test/java/ladder/service/LadderGameServiceTest.java
@@ -1,7 +1,7 @@
 package ladder.service;
 
 import ladder.domain.*;
-import ladder.service.dto.LadderGameResult;
+import ladder.controller.dto.LadderGameResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
